+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Command to import MNEMONIC or SEED wallet to encrypted file
+
 ## [1.1.0] - 2021-03-01
 
 ### Added